backend/requirements.txt
<<<<<<< HEAD
# Core FastAPI
fastapi==0.120.3
uvicorn[standard]==0.38.0
python-multipart==0.0.6
python-dotenv==1.0.0
starlette==0.49.1

# Authentication
PyJWT==2.8.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
cryptography==46.0.3

# Database
pymongo==4.15.3
=======
annotated-types==0.7.0
anyio==3.7.1
azure-cognitiveservices-speech==1.46.0
azure-core==1.36.0
bcrypt==4.0.1
boto3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.1.0
fastapi==0.120.3
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx<1,>=0.23.0
idna==3.11
jiter==0.11.1
>>>>>>> 70aa8405b58763a7ff5f09760677d9ea819d9595
motor==3.7.1
dnspython==2.8.0

# AWS
boto3==1.35.0

# Azure OpenAI
openai==2.6.1
azure-core==1.36.0

# HTTP
requests==2.31.0
httpx==0.28.1
httpcore==1.0.9

# Data validation
pydantic==2.12.3
pydantic_core==2.41.4
<<<<<<< HEAD
email-validator==2.1.0

# Utilities
PyYAML==6.0.3
click==8.3.0
tqdm==4.67.1
typing_extensions==4.15.0
=======
PyJWT==2.8.0
pymongo==4.15.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
sniffio==1.3.1
starlette==0.49.1
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
>>>>>>> 70aa8405b58763a7ff5f09760677d9ea819d9595
